Refactored RetryStrategy to be async
This refactors RetryStrategy to be async. This is necessary because
a strategy may need to use a synchronization primitive or may need
to internally wait.

For example, AWS retry strategies may need to wait internally even
if a retry would not be permitted if the operation is a long-poll
operation.

Strategies that use token buckets for client-side rate limiting may
also opt to wait for the token bucket to be refilled rather than
fast-failing when it is empty.
