+++ packages/smithy-aws-core/tests/unit/identity/test_imds.py
@@ -16,7 +16,7 @@ from smithy_aws_core.identity.imds import (
     TokenCache,
 )
 from smithy_core import URI
-from smithy_core.retries import SimpleRetryStrategy
+from smithy_core.aio.retries import SimpleRetryStrategy
 from smithy_http.aio import HTTPRequest
