+++ packages/smithy-core/tests/unit/aio/test_retries.py
@@ -1,0 +1,57 @@
+#  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#  SPDX-License-Identifier: Apache-2.0
+
+import pytest
+from smithy_core.aio.retries import SimpleRetryStrategy
+from smithy_core.exceptions import CallError, RetryError
+from smithy_core.retries import ExponentialRetryBackoffStrategy
+
+
+@pytest.mark.parametrize("max_attempts", [2, 3, 10])
+async def test_simple_retry_strategy(max_attempts: int) -> None:
+    strategy = SimpleRetryStrategy(
+        backoff_strategy=ExponentialRetryBackoffStrategy(backoff_scale_value=5),
+        max_attempts=max_attempts,
+    )
+    error = CallError(is_retry_safe=True)
+    token = await strategy.acquire_initial_retry_token()
+    for _ in range(max_attempts - 1):
+        token = await strategy.refresh_retry_token_for_retry(
+            token_to_renew=token, error=error
+        )
+    with pytest.raises(RetryError):
+        await strategy.refresh_retry_token_for_retry(token_to_renew=token, error=error)
+
+
+async def test_simple_retry_does_not_retry_unclassified() -> None:
+    strategy = SimpleRetryStrategy(
+        backoff_strategy=ExponentialRetryBackoffStrategy(backoff_scale_value=5),
+        max_attempts=2,
+    )
+    token = await strategy.acquire_initial_retry_token()
+    with pytest.raises(RetryError):
+        await strategy.refresh_retry_token_for_retry(
+            token_to_renew=token, error=Exception()
+        )
+
+
+async def test_simple_retry_does_not_retry_when_safety_unknown() -> None:
+    strategy = SimpleRetryStrategy(
+        backoff_strategy=ExponentialRetryBackoffStrategy(backoff_scale_value=5),
+        max_attempts=2,
+    )
+    error = CallError(is_retry_safe=None)
+    token = await strategy.acquire_initial_retry_token()
+    with pytest.raises(RetryError):
+        await strategy.refresh_retry_token_for_retry(token_to_renew=token, error=error)
+
+
+async def test_simple_retry_does_not_retry_unsafe() -> None:
+    strategy = SimpleRetryStrategy(
+        backoff_strategy=ExponentialRetryBackoffStrategy(backoff_scale_value=5),
+        max_attempts=2,
+    )
+    error = CallError(fault="client", is_retry_safe=False)
+    token = await strategy.acquire_initial_retry_token()
+    with pytest.raises(RetryError):
+        await strategy.refresh_retry_token_for_retry(token_to_renew=token, error=error)
